+++ src/Command/MessageInterface.php
@@ -55,7 +55,7 @@ interface MessageInterface
 
     const PROGRESS_DISPLAY = '%current%/%max% [%bar%] %percent:3s%% %elapsed:6s%/%estimated:-6s% Memory: %memory:6s%';
 
-    const PARSE_ERRORS = '<error>Parse Errors:</error>';
+    const PARSE_LOGS = 'Logs:';
 
     const NOTHING_TO_PARSE = '<error>No files found!</error>';