+++ tests/PhpDATest/Command/Strategy/InheritanceTest.php
@@ -105,8 +105,6 @@ class InheritanceTest extends \PHPUnit_Framework_TestCase
         $this->finder->shouldReceive('getIterator')->andReturn(array($file));
         $this->fixture->setOptions(array('output' => $this->output, 'config' => $this->config));
 
-        $this->collection->shouldReceive('hasAnalysisFailures')->andReturn(true);
-        $this->collection->shouldReceive('getAnalysisFailures')->andReturn(array('error' => new \Exception()));
         $this->analyzer->shouldReceive('analyze')->once()->with($file);
         $this->analyzer->shouldReceive('getAnalysisCollection')->andReturn($this->collection);
 
@@ -137,6 +135,10 @@ class InheritanceTest extends \PHPUnit_Framework_TestCase
         );
         $traverser->shouldReceive('bindVisitors')->once()->with(array(), $visitorOptions);
 
+        $logger = \Mockery::mock('PhpDA\Parser\Logger');
+        $logger->shouldReceive('isEmpty')->once()->andReturn(true);
+
         $this->analyzer->shouldReceive('getNodeTraverser')->once()->andReturn($traverser);
+        $this->analyzer->shouldReceive('getLogger')->once()->andReturn($logger);
     }
 }